# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  client.AlertHistoryPaginatedResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/client.AlertHistoryResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
      totalPages:
        type: integer
    type: object
  client.AlertHistoryResponse:
    properties:
      alertname:
        type: string
      annotations:
        additionalProperties: true
        type: object
      createdAt:
        type: string
      endsAt:
        type: string
      fingerprint:
        type: string
      labels:
        additionalProperties: true
        type: object
      landscape:
        type: string
      region:
        type: string
      severity:
        type: string
      startsAt:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
  client.ProjectsResponse:
    properties:
      projects:
        items:
          type: string
        type: array
    type: object
  client.UpdateLabelResponse:
    properties:
      fingerprint:
        type: string
      label:
        properties:
          key:
            type: string
          value:
            type: string
        type: object
      message:
        type: string
      project:
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: This is the backend API for the Developer Portal, providing endpoints
    for managing organizations, teams, projects, components, and deployments.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Developer Portal Backend API
  version: "1.0"
paths:
...
  /alert-history/alerts/{project}:
    get:
      consumes:
      - application/json
      description: Retrieve alerts for a specific project with optional filtering
        and pagination
      parameters:
      - description: Project name (must be one of cis2, usrv, cloud_automation)
        in: path
        name: project
        required: true
        type: string
      - default: 1
        description: Page number (min 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 50
        description: Items per page (min 1, max 100)
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - description: Filter by severity (e.g., critical, warning, info)
        in: query
        name: severity
        type: string
      - description: Filter by region label
        in: query
        name: region
        type: string
      - description: Filter by landscape
        in: query
        name: landscape
        type: string
      - description: Filter by status (firing or resolved)
        enum:
        - firing
        - resolved
        in: query
        name: status
        type: string
      - description: Filter by component label
        in: query
        name: component
        type: string
      - description: Filter by alert name
        in: query
        name: alertname
        type: string
      - description: Filter alerts after this time (RFC3339 format)
        in: query
        name: start_time
        type: string
      - description: Filter alerts before this time (RFC3339 format)
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved alerts
          schema:
            $ref: '#/definitions/client.AlertHistoryPaginatedResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get alerts by project
      tags:
      - alert-history
  /alert-history/alerts/{project}/{fingerprint}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific alert by its unique fingerprint within a project
      parameters:
      - description: Project name (alphanumeric, dashes, underscores)
        in: path
        name: project
        required: true
        type: string
      - description: Alert fingerprint (hexadecimal string, max 128 chars)
        in: path
        name: fingerprint
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved alert
          schema:
            $ref: '#/definitions/client.AlertHistoryResponse'
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Alert not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get alert by fingerprint
      tags:
      - alert-history
  /alert-history/alerts/{project}/{fingerprint}/label:
    put:
      consumes:
      - application/json
      description: Update or add a label for a specific alert
      parameters:
      - description: Project name (alphanumeric, dashes, underscores)
        in: path
        name: project
        required: true
        type: string
      - description: Alert fingerprint (hexadecimal string, max 128 chars)
        in: path
        name: fingerprint
        required: true
        type: string
      - description: Label update request with 'key' and 'value' fields
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Label updated successfully
          schema:
            $ref: '#/definitions/client.UpdateLabelResponse'
        "400":
          description: Invalid request parameters or body
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Alert not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update alert label
      tags:
      - alert-history
  /alert-history/projects:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all available projects configured in the alert
        history system
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved projects
          schema:
            $ref: '#/definitions/client.ProjectsResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all alert history projects
      tags:
      - alert-history
...
  /api/v1/projects/{projectId}/alerts:
    get:
      consumes:
      - application/json
      description: Fetches all Prometheus alert configurations from the configured
        GitHub repository
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Alerts data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get Prometheus alerts from GitHub repository
      tags:
      - alerts
  /api/v1/projects/{projectId}/alerts/pr:
    post:
      consumes:
      - application/json
      description: Creates a pull request to update Prometheus alert configurations
        in GitHub
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      - description: Alert changes
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: PR created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a pull request with alert changes
      tags:
      - alerts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"